common/types/form.tsのuser関連の型を移動
