+++ features/users/types/context.ts
@@ -1,5 +1,15 @@
-import { UserFormDataParams } from "@/common/types/form";
 import { Dispatch, SetStateAction } from "react";
+import { TeamFormDataParams } from "@/features/teams/types/context";
+
+export type UserFormDataParams = {
+  role: string;
+  name: string;
+  email: string;
+  profile_content: string;
+  password?: string;
+  password_confirmation?: string;
+  team: TeamFormDataParams;
+};
 
 export type UserFormContextProps = {
   userFormData: UserFormDataParams;
